Changed argument order in Inventory::takeItemSlot in order to fit standard order of importance. Simplified selling in Inventory::useShop and fixed bug with selling items to the shop. Also changed amount in struct Item to short as it should be.
git-svn-id: https://dev.chiasoft.net:8443/svn/svn_vana/trunk@668 23f3cdc0-e6d1-42be-a0c1-51f4125517bb

+++ src/ChannelServer/PlayerInventory.h
@@ -76,7 +76,7 @@ struct Item {
 		amount = item->amount;
 	}
 	int id;
-	int amount;
+	short amount;
 };
 typedef hash_map<short, Item *> iteminventory;
